style: clear the whole clippy backlog, and fix the one kernel it exposed
`cargo clippy` aborted in backend-cpu on a denied approx_constant lint until
af48cea, so it had never reached most of the workspace. With it unblocked the
real backlog was 123 files, not the 30 an earlier (cached) count suggested.

Mostly machine-applicable: is_multiple_of/div_ceil over manual modulo, needless
range loops rewritten as iterators, redundant same-type casts, repeat_n,
split_once, vec![] over push-after-new, and float literals trimmed to the digits
f32 actually keeps. Each rewritten Cephes constant in the AVX2 exp path was
checked to be BIT-IDENTICAL in f32 before accepting it (5.0000001201e-1 really
does round to 0.5 — it is 0.500000012, inside half a ULP), and the workspace

Four needed a judgment call rather than the suggested rewrite:
  * KernelVariant::from_str -> parse_str. It shadowed std::str::FromStr::from_str
    at the call site while returning Option rather than the trait's Result.
  * wgsl-cpu's to_i64 -> emit_i64: it appends an instruction to the function
    under construction, it does not convert self, so the to_* name was wrong.
  * data::toolcall's catalogue() return type became a named CatalogueEntry.
  * two doc-comment sites where a continuation line read as a new markdown list
    item, and one floating doc comment attached to no item.

`cargo build` and `cargo clippy --workspace --all-targets` are both at ZERO.

Separately, the clippy run made the wgsl-cpu JIT test reachable again, which
surfaced a REAL defect it had been hiding: max_abs_final was the one kernel of
346 that would not compile on the CPU backend, because it omitted the
num_workgroups argument the Cranelift JIT requires of every kernel. That meant
int8 quantization had no working CPU-backend path at all. It now declares the
argument and reconstructs the index the standard way; behaviour is unchanged
since it is always dispatched as a single workgroup.

Verified after the sweep: gradcheck 124 passed / 0 failed (including the four
new imaging gates), yolo, depth, vision, model, imaging, sam2, facenet, vqgan,
clip, capability, checkpoint all green, and all 346 kernels now compile on the
CPU JIT.

+++ crates/cli/src/imageops.rs
@@ -85,7 +85,7 @@ impl Action for Gradient {
         };
         let mut img = vec![0f32; wd * ht * 3];
         let (cx, cy) = (wd as f32 * 0.5, ht as f32 * 0.55);
-        let maxr = ((cx * cx + cy * cy) as f32).sqrt();
+        let maxr = (cx * cx + cy * cy).sqrt();
         for row in 0..ht {
             if row % 64 == 0 {
                 progress(Progress::step((row / 64) as u32 + 1, (ht / 64) as u32 + 1, "rendering"));
